se cambiar para que el input number también acepte decimales

+++ src/componentes/registrarCuota.js
@@ -82,7 +82,7 @@ function RegistrarCuotas() {
 
     const handleInputChange = (event) => {
         const { id, value } = event.target;
-        setFormulario({ ...formulario, [id]: parseInt(value.trim()) });
+        setFormulario({ ...formulario, [id]: parseFloat(value.trim()) });
     };
 
     const handleSubmit = async (event) => {
